refactor: Winner 클래스가 분리될 만큼의 역할을 하지 않는다고 판단하여 삭 해당 기능을 Cars에서 담당하도록 수정
