+++ src/main/java/racingcar/view/ResultView.java
@@ -4,11 +4,10 @@ import racingcar.domain.Car;
 import racingcar.domain.Cars;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class ResultView {
-    public static void printWinners(List<Car> winners) {
-        System.out.println(winners.stream().map(car -> car.getName().getName()).collect(Collectors.joining(",")) + "가 최종 우승했습니다.");
+    public static void printWinnerNames(List<String> winnerNames) {
+        System.out.println(String.join(",", winnerNames) + "가 최종 우승했습니다.");
     }
 
     public static void printRoundResult(Cars cars) {
